refactor: remove upload pkg and use internal
