+++ handlers/imagebbs/imagebbsBoardPage.go
@@ -23,7 +23,7 @@ import (
 	"github.com/disintegration/imaging"
 	"github.com/gorilla/mux"
 
-	"github.com/arran4/goa4web/pkg/upload"
+	"github.com/arran4/goa4web/internal/upload"
 	"github.com/arran4/goa4web/runtimeconfig"
 )
