+++ internal/startup/startup.go
@@ -7,7 +7,7 @@ import (
 
 	common "github.com/arran4/goa4web/core/common"
 	"github.com/arran4/goa4web/internal/dbstart"
-	"github.com/arran4/goa4web/pkg/upload"
+	"github.com/arran4/goa4web/internal/upload"
 	"github.com/arran4/goa4web/runtimeconfig"
 )
